AMBR-218 Move getThumbnailKey method to ArticleService
