+++ src/test/java/org/ambraproject/wombat/service/DataUriEncodeFunctionTest.java
@@ -1,13 +1,11 @@
 package org.ambraproject.wombat.service;
 
 import org.ambraproject.wombat.identity.RequestedDoiVersion;
-import org.ambraproject.wombat.service.remote.ContentKey;
 import org.apache.commons.io.IOUtils;
 import org.apache.http.Header;
 import org.apache.http.HeaderElement;
 import org.apache.http.HttpEntity;
 import org.apache.http.NameValuePair;
-import org.apache.http.entity.ContentType;
 import org.mockito.InjectMocks;
 import org.mockito.MockitoAnnotations;
 import org.testng.annotations.BeforeMethod;
@@ -24,11 +22,8 @@ import static org.junit.Assert.assertEquals;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.UUID;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -86,15 +81,5 @@ public class DataUriEncodeFunctionTest {
     
     assertEquals(Optional.of("data:image/x-foo;base64,YmFy"),
                  DataUriEncodeFunction.encodeAsDataUrl(entity));
-  }
-
-  @Test
-  public void testCreateKeyFromMap() {
-    Map<String, String> fileRepoMap = new HashMap<>();
-    fileRepoMap.put("crepoKey", "foo");
-    UUID uuid =  UUID.randomUUID();
-    fileRepoMap.put("crepoUuid", uuid.toString());
-    ContentKey key = DataUriEncodeFunction.createKeyFromMap(fileRepoMap);
-    assertEquals(String.format("[key: foo, uuid: %s]", uuid), key.toString());
   }
 }
